Todo hecho, solo falta comentar

+++ routes/users.js
@@ -3,6 +3,8 @@ const router = express.Router();
 const passport = require('passport');
 const user = require('../models/user');
 const Reserva = require('../models/reserva');
+const Producto = require('../models/producto');
+const Venta = require('../models/venta');
 const QRCode = require('qrcode');
 const bodyParser = require('body-parser');
 const fs = require('fs');
@@ -20,34 +22,54 @@ router.post('/signup', passport.authenticate('local-signup', {
 })); 
 
 router.get('/signin', (req, res, next) => {
-  res.render('login');
+  if (!req.user)
+    res.render('login');
+  else
+    res.redirect('/'); 
 });
 
 router.get('/validar', async (req, res, next) => {
-  const users= await user.find().sort({ email: 1});
-  res.render('validar', {users});
+  if (!req.user)
+    res.redirect('/');
+  else {
+    if (req.user.rol==="Administrador"){
+
+      const users= await user.find().sort({ email: 1});
+      res.render('validar', {users});
+
+    } else {
+      res.redirect('/');
+    }
+  }
 });
 
 router.post('/validar', async (req, res, next) => {
-  const { clientes, empleados } = req.body;
-  const clientesEmails = JSON.parse(clientes);
-  const empleadosEmails = JSON.parse(empleados);
+  if (!req.user)
+    res.redirect('/');
+  else{
+    if (req.user.rol==="Administrador"){
+      const { clientes, empleados } = req.body;
+      const clientesEmails = JSON.parse(clientes);
+      const empleadosEmails = JSON.parse(empleados);
 
-  const users= await user.find();
+      const users= await user.find();
 
-
-  for (const user of users){
-      if(empleadosEmails.includes(user.email)){
-          user.rol='Empleado';
-          await user.save();
+      for (const user of users){
+          if(empleadosEmails.includes(user.email)){
+              user.rol='Empleado';
+              await user.save();
+          }
+          else if(clientesEmails.includes(user.email)){
+              user.rol='Cliente';
+              await user.save();
+          }
       }
-      else if(clientesEmails.includes(user.email)){
-          user.rol='Cliente';
-          await user.save();
-      }
+      res.redirect('/catalogo');
+    }
+    else {
+      res.redirect('/');
+    }
   }
-
-  res.redirect('/catalogo');
 });
 
 router.post('/signin', passport.authenticate('local-signin', {
@@ -59,6 +81,7 @@ router.post('/signin', passport.authenticate('local-signin', {
 router.get('/profile',isAuthenticated, async (req, res, next) => {
   if (req.user!=null){
     let reservas = [];
+    let ventas = [];
     let hora="";
     if (req.user.rol==="Cliente"){
 
@@ -76,8 +99,11 @@ router.get('/profile',isAuthenticated, async (req, res, next) => {
         fecha: new Date(formattedDate), 
         cliente: req.user._id 
       }).sort({ hora: 1 });
-    }
 
+
+      ventas= await Venta.find({cliente: req.user._id}).sort({ fechaHora: -1 });
+    }else 
+      ventas= await Venta.find({empleado: req.user._id}).sort({ fechaHora: -1 });
 
     const dni= req.user.dni;
     
@@ -91,7 +117,7 @@ router.get('/profile',isAuthenticated, async (req, res, next) => {
       console.log('¡Código QR generado exitosamente!');
     });
 
-    res.render('profile',{qr: "/qr/qrcode"+ dni +".png", user: req.user, reservas, hora});
+    res.render('profile',{qr: "/qr/qrcode"+ dni +".png", user: req.user, reservas, hora, ventas});
 
   } else {
     res.redirect('/');
@@ -104,54 +130,61 @@ router.get('/logout', (req, res, next) => {
 });
 
 router.post('/editarUser', async (req, res) => {
-  const email = req.body.email;
+  if (!req.user)
+    res.redirect('/');
+  else{
+    const email = req.body.email;
 
-  try {
-    const editUser= await user.findOne({email : email});
+    try {
+      const editUser= await user.findOne({email : email});
 
-    if (!editUser) {
-      req.flash('error', 'El usuario no existe');
-      return res.redirect('/catalogo');
-    }else {
-      editUser.nombre=req.body.nombre;
-      editUser.apellidos=req.body.apellidos;
-      editUser.edad=req.body.edad;
+      if (!editUser) {
+        req.flash('error', 'El usuario no existe');
+        return res.redirect('/catalogo');
+      }else {
+        editUser.nombre=req.body.nombre;
+        editUser.apellidos=req.body.apellidos;
+        editUser.edad=req.body.edad;
 
-      await editUser.save();
+        await editUser.save();
+      }
+
+      res.redirect('/profile');
+    } catch (error) {
+      console.error('Error al modificar usuario:', error);
+      req.flash('error', 'Ha ocurrido un error al modificar el usuario');
+      res.redirect('/catalogo');
     }
-
-    res.redirect('/profile');
-  } catch (error) {
-    console.error('Error al modificar usuario:', error);
-    req.flash('error', 'Ha ocurrido un error al modificar el usuario');
-    res.redirect('/catalogo');
   }
 });
 
 router.post('/eliminarUser', isAuthenticated, async (req, res) => {
-
-  try {
-    const userId = req.body.userId;
-    const dni= req.user.dni;
-
-
-    //borrar QR
-    fs.unlink("public/qr/qrcode"+ dni +".png", async (err) => {});
+  if (!req.user)
+    res.redirect('/');
+  else{
+    try {
+      const userId = req.body.userId;
+      const dni= req.user.dni;
 
 
-    const reservas= await Reserva.find({cliente: userId});
-    for (const reserva of reservas){
-      await Reserva.findByIdAndDelete(reserva._id);
+      //borrar QR
+      fs.unlink("public/qr/qrcode"+ dni +".png", async (err) => {});
+
+
+      const reservas= await Reserva.find({cliente: userId});
+      for (const reserva of reservas){
+        await Reserva.findByIdAndDelete(reserva._id);
+      }
+
+      await user.findByIdAndDelete(userId);
+      res.redirect('/logout');
+
+    } catch (error) {
+      console.error('Error al eliminar usuario:', error);
+      req.flash('error', 'Ha ocurrido un error al eliminar el usuario');
+      // En caso de error, redirigir también a la página admin_profile.ejs
+      res.redirect('/profile');
     }
-
-    await user.findByIdAndDelete(userId);
-    res.redirect('/logout');
-
-  } catch (error) {
-    console.error('Error al eliminar usuario:', error);
-    req.flash('error', 'Ha ocurrido un error al eliminar el usuario');
-    // En caso de error, redirigir también a la página admin_profile.ejs
-    res.redirect('/profile');
   }
 });
